delete post related to user,fixing bug on delete user whitout photo

+++ app/Http/Controllers/AdminUsersController.php
@@ -131,7 +131,10 @@ class AdminUsersController extends Controller
     {
         //
         $user = User::findOrFail($id);
-        unlink(public_path() . $user->photo->file);
+        if ($user->photo()->exists()) {
+            unlink(public_path() . $user->photo->file);
+        }
+
         $user->delete();
         Session::flash('deleted_user', 'The user has beed deleted');
         return redirect('admin/users');
